Moved graph widget to seperate file
