+++ calc/calculator.kv
@@ -76,7 +76,7 @@
                 text:"Standard mode"
 
     GridLayout:
-        cols:2
+        cols:3
         pos:(0,root.height-75)
         width:300
         height:30
@@ -88,6 +88,9 @@
         Button:
             text:"Add function"
             on_press:root.add_function()
+        Button:
+            text:"Add variable"
+            on_press:root.add_vary_var()
 
     ScrollView:
         height:root.height-40-90-20-30
